+++ modules/functionapp/outputs-service-plan.tf
@@ -1,4 +1,0 @@
-output "app_service_plan_id" {
-  description = "ID of the created App Service Plan"
-  value       = var.app_service_plan_id
-}