+++ modules/windows-function/outputs-service-plan.tf
@@ -1,0 +1,4 @@
+output "service_plan_id" {
+  description = "Id of the created App Service Plan"
+  value       = var.service_plan_id
+}