+++ modules/windows-function/versions.tf
@@ -1,0 +1,14 @@
+terraform {
+  required_version = ">= 1.1"
+
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "~> 3.1"
+    }
+    azurecaf = {
+      source  = "aztfmod/azurecaf"
+      version = "~> 1.1"
+    }
+  }
+}