+++ outputs.tf
@@ -1,25 +1,9 @@
-output "function_app_id" {
-  description = "ID of the created Function App"
-  value       = module.function_app.function_app_id
+output "linux_function_app" {
+  description = "Linux Function App output object if Linux is choosen. Please refer to `./modules/linux-function/README.md`"
+  value       = try(module.linux_function["enabled"], null)
 }
 
-output "function_app_name" {
-  description = "Name of the created Function App"
-  value       = module.function_app.function_app_name
-}
-
-output "function_app_outbound_ip_addresses" {
-  description = "Outbound IP adresses of the created Function App"
-  value       = module.function_app.function_app_outbound_ip_addresses
-}
-
-output "function_app_connection_string" {
-  description = "Connection string of the created Function App"
-  value       = module.function_app.function_app_connection_string
-  sensitive   = true
-}
-
-output "function_app_identity" {
-  value       = module.function_app.function_app_identity
-  description = "Identity block output of the Function App"
+output "windows_function_app" {
+  description = "Windows Function App output object if Windows is choosen. Please refer to `./modules/windows-function/README.md`"
+  value       = try(module.windows_function["enabled"], null)
 }